[SNAP-2084] Handled dropStorageMemoryForObject in DefaultMemoryManager.
